% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golf_and_weather.R
\name{golf_map_weather}
\alias{golf_map_weather}
\title{Show golf courses with weather info on a map}
\usage{
golf_map_weather(api_key)
}
\arguments{
\item{api_key}{Your OpenWeatherMap API key.}
}
\description{
Show golf courses with weather info on a map
}
\examples{
\dontrun{
golf_map_weather(api_key = "your_api_key_here")
}
}
