% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golf_map.R
\name{show_golf_map}
\alias{show_golf_map}
\title{Show golf courses on a map}
\usage{
show_golf_map()
}
\description{
Show golf courses on a map
}
\details{
Opens a leaflet map with clickable markers for each golf course in Amsterdam.
}
\examples{
# Requires an interactive R session
show_golf_map()
}
