% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golf_and_weather.R
\name{start_golf_app}
\alias{start_golf_app}
\title{Run the golf app}
\usage{
start_golf_app()
}
\value{
A shiny popup of the Golf App; no arguments required
}
\description{
Run the golf app
}
\examples{
start_golf_app()
}
