% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golf_and_weather.R
\name{weather_by_coords}
\alias{weather_by_coords}
\title{Get current weather by coordinates}
\usage{
weather_by_coords(lat, lon, api_key)
}
\arguments{
\item{lat}{Latitude of the location.}

\item{lon}{Longitude of the location.}

\item{api_key}{OpenWeatherMap API key.}
}
\value{
Weather information of entered location.
}
\description{
Get current weather by coordinates
}
\examples{
\dontrun{
weather_by_coords(52.36409891117978, 4.911415030689269, "api_key_here")
}
}
